Fix JSON作成部分は外部から使用しないので分割した
