+++ pkg/qemu/install.go
@@ -2,12 +2,8 @@ package qemu
 
 import (
 	"bytes"
-	"encoding/json"
-	"fmt"
-	"os"
+	"github.com/CharVstack/CharV-lib/internal/qemu/install"
 	"text/template"
-
-	"github.com/google/uuid"
 )
 
 func CreateDisk(opts string) error {
@@ -24,77 +20,30 @@ func CreateDisk(opts string) error {
 	return run(cmd)
 }
 
-func CreateInfoJSON(opts InstallOpts) (Vm, error) {
-	uuidInt, err := uuid.NewRandom()
-	if err != nil {
-		return Vm{}, err
+func Install(opts InstallOpts) (install.Vm, error) {
+	requestData := install.RequestOpts{
+		Name:   opts.RequestData.Name,
+		Memory: opts.RequestData.Memory,
+		VCpu:   opts.RequestData.VCpu,
+		Image:  opts.RequestData.Image,
+		Disk:   opts.RequestData.Disk,
 	}
 
-	uuidString := uuidInt.String()
-
-	vmInfo := Vm{
-		Devices: Devices{
-			Disk: []*Disk{
-				{
-					Type: "qcow2",
-					Path: "/var/lib/charVstack/image/" + opts.Disk,
-				},
-			},
-		},
-		Memory: opts.Memory,
-		Metadata: Metadata{
-			ApiVersion: "v1",
-			Id:         uuidString,
-		},
-		Name: opts.Name,
-		VCpu: opts.VCpu,
-	}
-
-	var MarshalJSON []byte
-	MarshalJSON, err = json.Marshal(vmInfo)
-	if err != nil {
-		return Vm{}, err
-	}
-
-	createJSONPath := "/var/lib/charVstack/machines/"
-
-	fileName := createJSONPath + vmInfo.Name + "-" + vmInfo.Metadata.Id + ".json"
-
-	var createFile *os.File
-	createFile, err = os.Create(fileName)
-	if err != nil {
-		return Vm{}, err
-	}
-	defer func() {
-		err = createFile.Close()
-		if err != nil {
-			fmt.Println(err)
-		}
-	}()
-
-	_, err = createFile.Write(MarshalJSON)
-	if err != nil {
-		return Vm{}, err
-	}
-
-	return vmInfo, err
-}
-
-func Install(opts InstallOpts) (Vm, error) {
 	tmpl, err := template.New("install").Parse(`qemu-system-x86_64 -accel kvm -daemonize -display none -name guest={{.Name}} -smp {{.VCpu}} -m {{.Memory}} -cdrom /var/lib/charVstack/iso/{{.Image}} -boot order=d -drive file=/var/lib/charVstack/images/{{.Disk}}.qcow2,format=qcow2 -drive file=/var/lib/charVstack/bios/bios.bin,format=raw,if=pflash,readonly=on`)
 	if err != nil {
-		return Vm{}, err
+		return install.Vm{}, err
 	}
 	var buf bytes.Buffer
-	if err := tmpl.Execute(&buf, opts); err != nil {
-		return Vm{}, err
+	err = tmpl.Execute(&buf, requestData)
+	if err != nil {
+		return install.Vm{}, err
 	}
 	cmd := buf.String()
 
-	var getJSON Vm
-	getJSON, err = CreateInfoJSON(opts)
+	var getJSON install.Vm
+	getJSON, err = install.CreateInfoJSON(requestData)
 	if err != nil {
-		return Vm{}, err
+		return install.Vm{}, err
 	}
 
 	return getJSON, run(cmd)
