+++ pkg/qemu/type.go
@@ -1,42 +1,18 @@
 package qemu
 
-type Disk struct {
-	Type string `json:"type"`
-	Path string `json:"path"`
-}
-
-type Devices struct {
-	Disk []*Disk `json:"disk"`
-}
+import "github.com/CharVstack/CharV-lib/internal/qemu/install"
 
 type Machine struct {
-	Name    string  `json:"name"`
-	Memory  int     `json:"memory"`
-	VCpu    int     `json:"vcpu"`
-	Devices Devices `json:"devices"`
+	Name    string          `json:"name"`
+	Memory  int             `json:"memory"`
+	VCpu    int             `json:"vcpu"`
+	Devices install.Devices `json:"devices"`
 }
 
 type InstallOpts struct {
-	Name   string
-	Memory int
-	VCpu   int
-	Image  string
-	Disk   string
+	RequestData install.RequestOpts
 }
 
 type StartOpts struct {
 	Disk string
-}
-
-type Metadata struct {
-	ApiVersion string `json:"api_version"`
-	Id         string `json:"id"`
-}
-
-type Vm struct {
-	Devices  Devices  `json:"devices"`
-	Memory   int      `json:"memory"`
-	Metadata Metadata `json:"metadata"`
-	Name     string   `json:"name"`
-	VCpu     int      `json:"vcpu"`
 }
